requirements.txt
pillow
matplotlib
scikit-learn
numpy
pandas
streamlit